Moved the visualizer code to a fragment from the activity.
